Convert glossary to data
This follows the basic structure of definitions data that was used in
Roman Mosaics. Here though there are cross references such as “cross
database searching: See federated searching” which are identified as
`crossref` in the yaml and fed into the glossary template. Internal
linking has not been added at this point.