Optimize chapters for epub release
Change the filenames to use hypens between words for more consistent
links between chapters in the various formats.

This commit refactors the link_to helpers inside chapter content (as
opposed to places like the nav or sidebar) to use the chapter_link
helper which is defined in Canto. This outputs a standard link_to block
unless the epub environment is present, in which case the resulting
links are optimized for the Epub package (XHTML file extension, no
directory indexes, etc).

Finally, sort_order has been removed from the cover page for now.