Add cover page back to sort_order
This adds the index.html page back to the book as sort_order: 0.
However, the epub template already contains a cover.xhtml page (used to
show the book cover image). So I've changed the title of this page to
match the title of the book.

Finally, the button link for "introduction" on this page should not
appear in the epub version, so I've added a conditional block to keep it
from showing up in that build environment.