add replacement files and instructions for final ebook generation
In its current state, the automatically output epub needs a little work
to make it pass epubcheck-3.0.1 validation, to clean up the styling,
and to add a title page. These files assist in that. Once the work
outlined for the epub is complete, the mobi file can also be generated.
All instructions and notes are in the README documentation in this
directory.